requirements.txt
Django==1.8.1
Fabric==1.9.0
GitPython==0.1.7
amqp==1.4.6
anyjson==0.3.3
argparse==1.2.1
bcode==0.5
beautifulsoup4==4.3.2
billiard==3.3.0.20
celery==3.1.18
celery-haystack==0.8
django-appconf==1.0.1
django-celery==3.1.16
django-celery-transactions==0.3.1
django-debug-toolbar==1.2.1
django-haystack==2.3.1
ecdsa==0.11
elasticsearch==1.0.0
gunicorn==18.0
kombu==3.0.26
mechanize==0.2.5
paramiko==1.14.0
psycopg2==2.5.3
pycrypto==2.6.1
pyelasticsearch==0.6.1
python-dateutil==2.2
python-memcached==1.53
pytz==2015.2
redis==2.10.3
requests==2.3.0
selenium==2.46.0
simplejson==3.5.2
six==1.7.2
sqlparse==0.1.11
uWSGI==2.0.5.1
urllib3==1.8.2
wsgiref==0.1.2
